Bluetooth: Controller: Fix connection update window offset data type
Fix regression in refactored LLCP using uint16_t instead of
uint32_t in storing the win_offset_us value. This caused
connection update to fail with incorrect window offset being
used to schedule the connection radio events.

Regression since commit 32cdfdd29667 ("Bluetooth: controller:
llcp: set refactored as default").

+++ subsys/bluetooth/controller/ll_sw/ull_conn.c
@@ -2088,7 +2088,7 @@ static void ull_conn_update_ticker(struct ll_conn *conn,
 }
 
 void ull_conn_update_parameters(struct ll_conn *conn, uint8_t is_cu_proc, uint8_t win_size,
-				uint16_t win_offset_us, uint16_t interval, uint16_t latency,
+				uint32_t win_offset_us, uint16_t interval, uint16_t latency,
 				uint16_t timeout, uint16_t instant)
 {
 	struct lll_conn *lll;
